Add explanation to use images in scatter
As a beginner following the example: image scatter. I would have like to know how to load any image, not Makie.logo(). Thus, I have added comments to explain that in the code. I think Makie.logo() is not representative of most image scatter that a user would want to create