Add explanation to use images in scatter
As a beginner following the example: image scatter. I would have like to know how to load any image, not Makie.logo(). Thus, I have added comments to explain that in the code. I think Makie.logo() is not representative of most image scatter that a user would want to create

Former-commit-id: ad82918df4f3b96967f23ba8a1ff9abc0ebb8798