Clean up and fix antialias scaling implementation
A big cleanup to the antialiasing scale calculation now that I've
understood it better. The idea is to compute the isotropic *scale
factor* of the uv->viewport transformation, at the centrepoint of the
glyph.

The key fix here was to incorporate the differential of the perspective
divide into the scale calculation, making sure it's evaluated at the
glyph centre.

Also fix the scale factor for sampled signed distance functions of
glyphs by using the average of x and y pixels per uv step. It's unclear
why these wouldn't be equivalent...