Test on Gitlab and implement "OpenGL-aware" testing (#152)
Creates a `gitlab-ci.yml` which allows Gitlab to test AbstractPlotting in the same way as Makie.  

If a GPU is detected on the system, then it will test using the GPU; otherwise, it will test software-only.  Gitlab caching has also been implemented, so the results of changes can be seen per commit.

Cache interval is 2 days, after which cached content is dropped.

Former-commit-id: 4611b211a0cd02539b84bd4000990d506b45f1c7