first working version for CairoMakie
this fixes wonky glyphs when using annotations! but not simple text! for some reason. also in GLMakie it seems to introduce some slight bugs with minus dashes of tick labels. also I thought that the conversion factor for freetype units was 64, but it seems I have to additionally divide the resulting glyph positions by 4 in order to make it work

Former-commit-id: b0e2c11a991d60b3b3c7bd49c23328cc2f594f73