give poly default strokewidth of 1
If you want to give poly a stroke, you always have to set both strokecolor and strokewidth right now which is quite annoying. I think if you specify a color then a stroke in that color should be visible, then you can still tweak the linewidth but 1 is a good default. It could also be the other way around, but I think it's more natural to think about the color first.

This won't change any plot using only the defaults, because the effect of strokewidth 1 and color = transparent is still an invisible stroke.