Allow colorbuffer to be used directly on Scenes
This is something I came across recently as a usecase - of course, it's trivial to do this myself, but I figured it should be in the docs.  I actually need to take the Fourier transform of a plot (it's a simulated worm waveform) and I could see someone having this need in the future.

We should also document this in the Gallery.

Former-commit-id: 2071e9ee75a92385458422e4eb3c690fd2dcabfa