improve slider drag behavior
move the button to the place where the first mousedown in the scene happens, so you can start sliding from anywhere. also tie the button color change to mousedown / up of the scene. basically, the button serves no special dragging function anymore, it's just an indicator of the slider position. this should make the slider much more robust and cause fewer glitches