fix bug in mousestatemachine that sometimes prohibited draggin
The problem depended on the order in which mouse movement and drag events came in. In the wrong order, `last_mouseevent` was set to `pressed` before the logic for `down` started a drag. Now, only up and down are allowed in `last_mouseevent` so the dragging logic is not skipped.