Fix identifier conflict warning on package load
Currently, if you load AbstractPlotting before CairoMakie, you get a
warning that CairoMakie.AbstractPlotting conflicts with an existing
identifier. This seems to happen due to the hard vs. soft binding
distinction in `import` vs. `using`, respectively. If you swap the
`import` for `using` when reexporting AbstractPlotting's symbols, the
result is the same, but the warning disappears.