do not export `step!`
This is more like a discussion rather than a fix. In Slack I've raised the point that several Julia packages export step. The ones I know are DifferentialEquations.jl, DynamicalSystems.jl and Agents.jl. 

I'm not sure if the exported `step!` here is used somewhere, but I don't know either.