Implement inherit for nested theme attributes
This allows plots to inherit from nested attrs in the scene's theme, 
i.e., a plot can inherit `scene.Axis.yticklabelsvisible` if it exists.

This is arbitrary and robust, following the same logic (check if exists 
in current scene, if not go to parent, if not return default value) as 
the implementation for a single Symbol did.