Rework shading for partly hidden elements
We now put a hatched shade on top of a legend entry if at least one of
its associated plot elements is hidden. If all are hidden we put
a full shade.
Also removed middle mouse click and made the right click such that
it reset the visibility for all plot elements.