Make GLMakie aware of the content scale factor on HiDPI screens
This is accomplished by making the window and rendered framebuffer
sizes independent of one another. By separating the two concepts, we
can create rasterized graphics at higher (or lower) resolution than
what is shown on screen.

The window scale factor is used to scale the sizes of elements within
the window (and on Linux and Windows, the window size itself from the
requested logical sizes), and the px-per-unit scale factor dictates
the size of the rasterized render.

The default scaling factors are initialized by requesting the content
scale factor from GLFW (and dynamically responding to changes in the
content scale factor on platforms where GLFW provides support).

Testing is added on Linux where xvfb-run can be used to force a
HiDPI context.