Make window and rendering framebuffer sizes independent
By separating the two concepts, we can create rasterized graphics at
higher resolution than what is shown on screen.

The window scale factor is used to scale the sizes of elements within
the window (and on Linux and Windows, the window size itself from the
requested logical sizes), and the px-per-unit scale factor dictates
the size of the rasterized render.