Make `get_all_plots` in CairoMakie fully flatten plots
I have a recipe which plots multiple polys, and translates them based on a parameter.  The idea is that only one poly has color, the rest only have stroke and the color is transparent, and they should all be seen as an overlay.

It turns out that CairoMakie only orders by the z-value of plots in `scene.plots`, so any translation within a recipe (or by a user of a recipe's plots) is not respected.

This commit fixes that by making `get_all_plots` truly recursive.  Now, it only stops if a dispatch is defined on a particular plot type, or if `plot.children` is empty, in which case it's treated as an atomic plot.

This seems to produce the correct results for my recipe.