Sf/vscode resizing (#3044)
* [WGLMakie] Add automatic figure resizing

This implements basic resizing support for WGLMakie figures.  This has
primarily been tested in the VSCode plot pane and the JSServe
auto-launched browser window (using Chrome and Safari).

The implementation creates a new `resize` event that the javascript
frontend can send back to the Julia backend, with a computed size (in
pixels) that the backend should resize to.  Additionally, the javascript
`throttle()` function is altered to no longer drop the last few events
when throttled, it instead delays the last event until the end of the
throttling period, which is important when resizing quickly so that the
final size is faithfully recorded.

* `deno bundle wglmakie.js`

* make resizing to body themeable

* update bundle

* fix fallout from removing window.WGLMakie = ...

* fix aspect ratio of Cam3D

* add news entry

---------

Co-authored-by: Elliot Saba <staticfloat@gmail.com>