DataInspector: fix an attribute extraction bug for heatmaps & images
If the colorrange attribute is left at default it returns
MakieCore.Automatic; the method for `get` that was called at
show_imagelike does not check for this type so the ensuing call to
interpolated_getindex would fail.

The colorrange is now generically extracted and
calculated_colors[].colorrange[] is preferred as it contains the actual
range.

Fixes #3101