Histogram: fix a bug when data changes to allequal values
When plotting Observables if the contained collection changed such that
all values were equal the minimum and maximum values would match causing
pick_hist_edges to return a vector. This in turn would result in a
conversion error if the previous value was a range. This is fixed by
simply returning a range instead.

Fixes #2452