Fix a DataInspector bug if `inspector_label` is used with RGB images
If the creation of a custom tooltip was attempted with RGB type images
the inspector would error after attempting to convert the Colorant to a
Float32 using the Point3f constructor for the position parameter.

For cases in which RGB values are being plotted we now call the
inspector_label function using a Tuple for its third argument instead,
avoiding the error and allowing the RGB values to be used in the
construction of the custom string.