Fix a hang due to `limits!` being called recursively
Fixes #3564

Restricting the sans axis arguments to the expected types prevents the
method from being called in a non-terminating loop which was the case
if, for example, the function was called with an indexed Figure instead
of an Axis; calling limits! without a current axis also resulted in a
stack overflow.