Check converted as well as input args in polygon drawing in CairoMakie (#3605)
CairoMakie decided whether or not to draw polygons as meshes based on the type of the input arguments. However, this excluded types which are later converted to polygons via convert_arguments from the direct polygon drawing path, forcing them to be meshed and rasterized.

This PR enables CairoMakie to decide whether to use meshes or not based on both the input and converted argument types. It first checks whether there is an available draw_poly method for the input types (plot.args...), if there is then it proceeds. If not, it checks whether there is an available draw_poly method for the converted types (plot.converted...), and proceeds there. The last/base case is of course a call to draw_poly_as_mesh.

This has eliminated the generic fallback for draw_poly(scene, screen, plot) as the decision-making is now done in draw_plot.

As an example: Rect, BezierPath, Vector{<: Point}, or Polygon input will be processed directly. Inputs from other geometry packages like ArchGDAL, GeoJSON, etc which were previously interpreted as requiring meshing will now be drawn as polygons, since after conversion they are usually GeometryBasics.Polygon or similar which have dispatches in draw_poly.