Ensure that the axis type is not an Observable when passed explicitly
If plotting, e.g. 
```julia
heatmap(rand(10, 10); axis = (; type = LScene))
```
I ran into the error that the parameter `AxType` sent to `set_axis_attributes!` and `_block` was an Observable.  This PR fixes that by simply ensuring that whatever is passed to the functions is always a value, not an Observable.