updated JS OrbitControls to match Julia OrbitControls
In three.js, the `.up` attribute of the camera is used as the rotation axis of
the OrbitControls. To match the Julia defaults, which rotate about the z-axis,
`.up` should be initialized to [0, 0, 1]. Also, `.target` (and `.target0`)
should be set to the correct point.

The OrbitControls are now attached to the three.js scene, allowing later
reference to them.