Follow up #4062, fix missed changes in merge commits
In PR #4062, the original changes were merged with `master` to fix
conflicts (presumably from #4073, it seems). During the merge
resolution, though, it seems a mistake was made and setting the
`GLFW.FLOATING` attribute was left in an unmodified state. Fix that
error.

At the same time, re-include some comment changes that also appear to
have gone missing.