GLMakie: fix volume.frag showing objects from behind the camera (#5164)
* GLMakie: fix volume.frag showing objects from behind the camera

This is a fix for https://github.com/MakieOrg/Makie.jl/issues/5085.

To test/reproduce this bug:

    using GLMakie
    vol = rand(1000,50,50);
    fig, scene, _ = volume(vol)

    controls = cameracontrols(scene)

    # position the camera at the long end:
    controls.eyeposition[] = Vec3f(1300, 25, 25)
    controls.lookat[] = Point3f(1000, 25, 25)
    update_cam!(scene.scene)

Next, press w to enter the volume. Once inside notice how elements from behind
seem to get mirrored across the direction of view, creating an effect where
part of the volume is moving past you from behind.

* GLMakie: fix order of start and stop positions in volume.frag

The mip algorithm does not care about the direction in which the ray is
traversed. Other algorithms, however, do.

* port changes to WGLMakie

* add refimg

* fix test

* fix format

* update changelog

---------

Co-authored-by: ffreyer <frederic481994@hotmail.de>